Move test of precess rx to the integration tests since it requires rx. Simplify some of the ray_tracing tests by using fixtures
