+++ tests/unit_tests/test_ray_tracing.py
@@ -1,6 +1,5 @@
 #!/usr/bin/python3
 
-import gzip
 import logging
 import shutil
 
@@ -144,36 +143,6 @@ def test_ray_tracing_read_results(ray_tracing_lst):
     assert ray_tracing_lst._has_results is True
     assert len(ray_tracing_lst._results) > 0
     assert ray_tracing_lst.get_mean("d80_cm").value == pytest.approx(4.256768651160611, abs=1e-5)
-
-
-def test_process_rx(
-    simtel_path_no_mock, io_handler, telescope_model_lst, tmp_test_directory, caplog
-):
-    """
-    Test the process_rx method of the RayTracing class with an empty file
-    and a non-existing file
-    """
-
-    config_data = {
-        "source_distance": 10 * u.km,
-        "zenith_angle": 20 * u.deg,
-        "off_axis_angle": [0, 0] * u.deg,
-    }
-
-    ray = RayTracing(
-        telescope_model=telescope_model_lst,
-        simtel_source_path=simtel_path_no_mock,
-        config_data=config_data,
-        label="empty_file",
-    )
-    with gzip.open(tmp_test_directory / "empty_file.gz", "wb"):
-        pass
-    with pytest.raises(IndexError):
-        ray._process_rx(file=tmp_test_directory / "empty_file.gz")
-        assert "Invalid output from rx" in caplog.text
-    with pytest.raises(FileNotFoundError):
-        ray._process_rx(file=tmp_test_directory / "non_existing_file.gz")
-        assert "Photon list file not found" in caplog.text
 
 
 def test_export_results(simtel_path, telescope_model_lst, caplog):
